requirements.txt
tensorflow
tensorflow-datasets
numpy
matplotlib
seaborn
pandas
pillow
flask
gunicorn
scikit-learn
